backend/requirements.txt
fastapi
uvicorn
openai
pydantic[email]
pydantic-settings
python-dotenv
sse-starlette
httpx
rich
playwright>=1.42.0
markdownify
docker
websockets
motor>=3.3.2
pymongo>=4.6.1
beanie>=1.25.0
async-lru>=2.0.0
redis>=5.0.1
python-jose[cryptography]
passlib[bcrypt]
python-multipart